+++ src/main/java/wooteco/subway/line/LineService.java
@@ -6,6 +6,7 @@ import wooteco.subway.line.dto.LineRequest;
 import wooteco.subway.line.dto.LineResponse;
 import wooteco.subway.line.exception.LineError;
 import wooteco.subway.line.exception.LineException;
+import wooteco.subway.section.*;
 import wooteco.subway.station.Station;
 import wooteco.subway.station.StationDao;
 import wooteco.subway.station.exception.StationError;
